Reviewer comment: turn into method, parameters
As suggested, turn this function into a method on the net. Furthermore,
remove the options to preserve the criteria and the history.
