+++ skorch/tests/test_utils.py
@@ -1,12 +1,10 @@
 """Test for utils.py"""
 
-import pickle
 from copy import deepcopy
 
 import numpy as np
 import pytest
 from scipy import sparse
-from sklearn.base import clone
 import torch
 from torch.nn.utils.rnn import PackedSequence
 from torch.nn.utils.rnn import pack_padded_sequence
@@ -876,154 +874,3 @@ class TestInferPredictNonlinearity:
         X = np.random.random((20, 5))
         out = fn(X)
         assert out is X
-
-
-class TestTrimForPrediction:
-    @pytest.fixture
-    def trim_for_prediction(self):
-        from skorch.utils import trim_for_prediction
-
-        return trim_for_prediction
-
-    @pytest.fixture
-    def net_untrained(self, classifier_module):
-        """A net with a custom 'module2_' criterion and a progress bar callback"""
-        from skorch import NeuralNetClassifier
-        from skorch.callbacks import ProgressBar
-
-        net = NeuralNetClassifier(
-            classifier_module,
-            max_epochs=2,
-            callbacks=[ProgressBar()],
-        )
-        return net
-
-    @pytest.fixture
-    def net(self, net_untrained, classifier_data):
-        X, y = classifier_data
-        return net_untrained.fit(X[:100], y[:100])
-
-    @pytest.fixture
-    def net_2_criteria(self, classifier_module, classifier_data):
-        """A net with a custom 'module2_' criterion and disabled callbacks"""
-        # Check that not only the standard components are trimmed and that
-        # callbacks don't need to be lists.
-
-        from skorch import NeuralNetClassifier
-
-        class MyNet(NeuralNetClassifier):
-            def initialize_criterion(self):
-                super().initialize_criterion()
-                # pylint: disable=attribute-defined-outside-init
-                self.criterion2_ = classifier_module()
-                return self
-
-        X, y = classifier_data
-        net = MyNet(classifier_module,  max_epochs=2, callbacks='disable')
-        net.fit(X, y)
-        return net
-
-    def test_trimmed_net_less_memory(self, net, trim_for_prediction):
-        # very rough way of checking for smaller memory footprint
-        size_before = len(pickle.dumps(net))
-        trim_for_prediction(net)
-        size_after = len(pickle.dumps(net))
-        # check if there is at least 10% size gain
-        assert 0.9 * size_before > size_after
-
-    def test_trim_untrained_net_raises(self, net_untrained, trim_for_prediction):
-        from skorch.exceptions import NotInitializedError
-
-        with pytest.raises(NotInitializedError):
-            trim_for_prediction(net_untrained)
-
-    def test_try_trimmed_net_fit_raises(self, net, trim_for_prediction, classifier_data):
-        from skorch.exceptions import SkorchTrainingImpossibleError
-
-        X, y = classifier_data
-        msg = (
-            "The net's attributes were trimmed for prediction, thus it cannot "
-            "be used for training anymore")
-
-        trim_for_prediction(net)
-        with pytest.raises(SkorchTrainingImpossibleError, match=msg):
-            net.fit(X, y)
-
-    def test_try_trimmed_net_partial_fit_raises(
-            self, net, trim_for_prediction, classifier_data
-    ):
-        from skorch.exceptions import SkorchTrainingImpossibleError
-
-        X, y = classifier_data
-        msg = (
-            "The net's attributes were trimmed for prediction, thus it cannot "
-            "be used for training anymore"
-        )
-
-        trim_for_prediction(net)
-        with pytest.raises(SkorchTrainingImpossibleError, match=msg):
-            net.partial_fit(X, y)
-
-    def test_inference_works(self, net, trim_for_prediction, classifier_data):
-        # does not raise
-        trim_for_prediction(net)
-        X, _ = classifier_data
-        net.predict(X)
-        net.predict_proba(X)
-        net.forward(X)
-
-    def test_trim_twice_works(self, net, trim_for_prediction):
-        # does not raise
-        trim_for_prediction(net)
-        trim_for_prediction(net)
-
-    def test_callbacks_trimmed(self, net, trim_for_prediction):
-        trim_for_prediction(net)
-        assert not net.callbacks
-        assert not net.callbacks_
-
-    def test_optimizer_trimmed(self, net, trim_for_prediction):
-        trim_for_prediction(net)
-        assert net.optimizer is None
-        assert net.optimizer_ is None
-
-    def test_criteria_trimmed(self, net_2_criteria, trim_for_prediction):
-        trim_for_prediction(net_2_criteria)
-        assert net_2_criteria.criterion is None
-        assert net_2_criteria.criterion_ is None
-        assert net_2_criteria.criterion2_ is None
-
-    def test_criteria_left_untrimmed(self, net_2_criteria, trim_for_prediction):
-        trim_for_prediction(net_2_criteria, trim_criterion=False)
-        assert net_2_criteria.criterion is not None
-        assert net_2_criteria.criterion_ is not None
-        assert net_2_criteria.criterion2_ is not None
-
-    def test_history_trimmed(self, net, trim_for_prediction):
-        trim_for_prediction(net)
-        assert not net.history
-
-    def test_history_left_untrimmed(self, net, trim_for_prediction):
-        trim_for_prediction(net, trim_history=False)
-        assert net.history
-
-    def test_module_training(self, net, trim_for_prediction):
-        # pylint: disable=protected-access
-        net._set_training(True)
-        trim_for_prediction(net)
-        assert net.module_.training is False
-
-    def test_can_be_pickled(self, net, trim_for_prediction):
-        pickle.dumps(net)
-        trim_for_prediction(net)
-        pickle.dumps(net)
-
-    def test_can_be_copied(self, net, trim_for_prediction):
-        deepcopy(net)
-        trim_for_prediction(net)
-        deepcopy(net)
-
-    def test_can_be_cloned(self, net, trim_for_prediction):
-        clone(net)
-        trim_for_prediction(net)
-        clone(net)
